data/dataset.py
import os
import json
import torch
from torch.utils.data import Dataset, DataLoader
from PIL import Image
from transformers import AutoTokenizer

class StepCoTDataset(Dataset):
    """Dataset for Step-CoT."""
    def __init__(self, json_path, image_root, tokenizer, max_len=128, transform=None):

        with open(json_path, 'r') as f:
            self.data = json.load(f)  # list of dicts, each with patient_id, image_path, report, vqa_chain

    def __len__(self):
        return len(self.data)

    def __getitem__(self, idx):

        return {
            "images": image,
            "labels": labels,
            "valid_mask": valid_mask,
            "tokenized_prompts": tokenized_prompts,
            "patient_id": item["patient_id"],
            "report": item["report"],
            "vqa_chain": vqa_chain
        }

def make_collate_fn(step_num_classes, max_length, tokenizer):
    """Collate function that stacks batches."""
    def collate_fn(batch):
        batch_size = len(batch)
        images = torch.stack([item["images"] for item in batch])
        labels = torch.stack([item["labels"] for item in batch])   # (B, S)

    return collate_fn

def get_dataloaders(config):
    """User must implement this to return train_loader, val_loader, test_loader."""
    # TODO: Replace with actual data loading using StepCoTDataset
    # Example:
    # tokenizer = AutoTokenizer.from_pretrained(config['teacher']['text_model_name'])
    # train_dataset = StepCoTDataset('path/to/train.json', config['data_root'], tokenizer, config['max_len'])
    # train_loader = DataLoader(train_dataset, batch_size=config['batch_size'], shuffle=True,
    #                           collate_fn=make_collate_fn(config['step_num_classes'], config['max_len'], tokenizer))
    # ...
    pass